Move renamedt class to its own file
Make navigating the classes easier.
